requirements.txt
Django==2.2.19
pytz==2021.1
sqlparse==0.4.1 
django-filter==22.1
djangorestframework==3.12.4